requirements.txt
nemo_toolkit[asr]==2.6.1
soundfile
librosa

# torch and torchaudio must be installed BEFORE this file.
# Install with CUDA support (recommended):
#   pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu121
# Or CPU-only:
#   pip install torch torchaudio --index-url https://download.pytorch.org/whl/cpu
#
# Note: If you encounter "object.__init__() takes exactly one argument" error,
# run the post-install fix: python fix_lhotse.py
